bcache: Fix 32 bit build
note: NUM_GC_GENS is now 8

+++ drivers/md/bcache/ioctl.c
@@ -243,7 +243,8 @@ static int bch_ioctl_list_keys_fn(struct btree_op *b_op, struct btree *b,
 	    op->i.buf_size / sizeof(u64))
 		return -ENOSPC;
 
-	if (copy_to_user((u64 __user *) op->i.buf + op->i.keys_found,
+	if (copy_to_user(((u64 __user *) (unsigned long)
+			  op->i.buf) + op->i.keys_found,
 			 k, bkey_bytes(k)))
 		return -EFAULT;
